+++ src/middleware/index.ts
@@ -1,5 +1,4 @@
-import type { RequestLog } from './logger';
 import { validate } from './validate';
 import { logger } from './logger';
 
-export { RequestLog, validate, logger };
+export { validate, logger };
